Code Cleanup. Renaming and moving class to new file.
